Update FileUtil.java (#24)
* Update FileUtil.java

Update to the FileUtil to make the java script resource loading work in an OSGI environment.   Kept the original loader to ensure backwards compatibility.  Essentially checks to make sure the InputStream is found.   If not, it will try loading the resource using the proper method for an OSGI environment.   If it still is unable to load the file, it will throw an exception to avoid a non descriptive null pointer exception.

* Update FileUtil.java

Forgot the import java.io.InputStream